requirements.txt
aiogram==2.25.1
aiogram-calendar==0.1.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
gino==1.0.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
peewee==3.15.4
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
SQLAlchemy==1.3.24
urllib3==1.26.14
yarl==1.8.2
